refactor: Move footer rendering logic and getArrayStatus function to a new Footer.php file for improved organization and maintainability.
